backend/requirements.txt
asgiref
Django
Django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
pytest-django
gunicorn
celery
django-celery-beat
redis
Pillow